+++ packages/blobs/src/environment.ts
@@ -47,7 +47,7 @@ export const getEnvironment = (): EnvironmentVariables => {
 declare global {
   // Using `var` so that the declaration is hoisted in such a way that we can
   // reference it before it's initialized.
-  // eslint-disable-next-line no-var
+
   var netlifyBlobsContext: unknown
 }