+++ packages/runtime/src/lib/globals.ts
@@ -7,7 +7,6 @@ declare global {
   // Using `var` so that the declaration is hoisted in such a way that we can
   // reference it before it's initialized.
 
-  // eslint-disable-next-line no-var
   var Netlify: NetlifyGlobal
 }